stt-server/requirements.txt
websockets
openai-whisper
numpy
torch
python-dotenv